🎯 moved `thebe` providers to `jupyter`
