+++ packages/jupyter/src/index.tsx
@@ -4,4 +4,6 @@ const OUTPUT_RENDERERS = {
   output: Output,
 };
 
+export * from './providers';
+
 export default OUTPUT_RENDERERS;
