+++ packages/jupyter/src/output.tsx
@@ -4,7 +4,7 @@ import type { MinifiedMimeOutput, MinifiedOutput } from 'nbtx';
 import classNames from 'classnames';
 import { SafeOutputs } from './safe';
 import { JupyterOutputs } from './jupyter';
-import { useNotebookCellExecution } from '@myst-theme/providers';
+import { useNotebookCellExecution } from './providers';
 
 export const DIRECT_OUTPUT_TYPES = new Set(['stream', 'error']);
